backend/requirements.txt
# This file lists the Python dependencies for the backend of the project.
fastapi
uvicorn
pydantic
pymongo
pytest
pytest-asyncio
httpx
python-dotenv
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
passlib[argon2]
python-multipart
